feat(correct json for postgres): Other features in media res
